Return the collection of available categories as a sorted list.
This is very much a "works for me" kinda patch. We need this so that the
Category "order" property is actually obeyed. I'm not sure if this really the
best way of doing this, but I don't think swapping a set for a list
should break anything.

Without this we get into the odd state that tshirts can come before
child support.